+++ mvc/src/main/java/com/interface21/webmvc/servlet/mvc/tobe/ControllerScanner.java
@@ -1,0 +1,35 @@
+package com.interface21.webmvc.servlet.mvc.tobe;
+
+import com.interface21.context.stereotype.Controller;
+import java.lang.reflect.InvocationTargetException;
+import java.util.HashMap;
+import java.util.Map;
+import java.util.Set;
+import org.reflections.Reflections;
+
+public class ControllerScanner {
+
+    public Map<Class<?>, Object> scan(Object... basePackage) {
+        Reflections reflections = new Reflections(basePackage);
+        Set<Class<?>> controllerClasses = reflections.getTypesAnnotatedWith(Controller.class);
+
+        final Map<Class<?>, Object> controllers = new HashMap<>();
+
+        for (Class<?> controllerClass : controllerClasses) {
+            controllers.put(controllerClass, createController(controllerClass));
+        }
+
+        return controllers;
+    }
+
+    private Object createController(final Class<?> controllerClass) {
+        Object controllerInstance;
+        try {
+            controllerInstance = controllerClass.getDeclaredConstructor().newInstance();
+        } catch (InstantiationException | IllegalAccessException | InvocationTargetException |
+                 NoSuchMethodException e) {
+            throw new RuntimeException(e);
+        }
+        return controllerInstance;
+    }
+}
